change: Extract TaskToolBarIcons from TaskListPage and use TypeScript
