+++ src/web/pages/tasks/ListPage.jsx
@@ -6,22 +6,17 @@
 import React from 'react';
 import {TASKS_FILTER_FILTER} from 'gmp/models/filter';
 import DashboardControls from 'web/components/dashboard/Controls';
-import {TaskIcon, WizardIcon} from 'web/components/icon';
-import ManualIcon from 'web/components/icon/ManualIcon';
-import IconDivider from 'web/components/layout/IconDivider';
+import {TaskIcon} from 'web/components/icon';
 import PageTitle from 'web/components/layout/PageTitle';
 import {
   USE_DEFAULT_RELOAD_INTERVAL,
   USE_DEFAULT_RELOAD_INTERVAL_ACTIVE,
 } from 'web/components/loading/Reload';
-import IconMenu from 'web/components/menu/IconMenu';
-import MenuEntry from 'web/components/menu/MenuEntry';
 import EntitiesPage from 'web/entities/Page';
 import withEntitiesContainer from 'web/entities/withEntitiesContainer';
-import useCapabilities from 'web/hooks/useCapabilities';
 import useTranslation from 'web/hooks/useTranslation';
 import TaskDashboard, {TASK_DASHBOARD_ID} from 'web/pages/tasks/dashboard';
-import NewIconMenu from 'web/pages/tasks/icons/NewIconMenu';
+import TaskToolBarIcons from 'web/pages/tasks/icons/TaskToolBarIcons';
 import TaskComponent from 'web/pages/tasks/TaskComponentComponent';
 import TaskFilterDialog from 'web/pages/tasks/TaskFilterDialog';
 import TaskTable from 'web/pages/tasks/TaskTable';
@@ -30,67 +25,6 @@ import {
   selector as entitiesSelector,
 } from 'web/store/entities/tasks';
 import PropTypes from 'web/utils/PropTypes';
-
-export const ToolBarIcons = ({
-  onAdvancedTaskWizardClick,
-  onModifyTaskWizardClick,
-  onContainerTaskCreateClick,
-  onTaskCreateClick,
-  onTaskWizardClick,
-}) => {
-  const capabilities = useCapabilities();
-  const [_] = useTranslation();
-  const mayUseModifyTaskWizard =
-    capabilities.mayEdit('task') &&
-    (capabilities.mayCreate('alert') || capabilities.mayCreate('schedule'));
-  return (
-    <IconDivider>
-      <ManualIcon
-        anchor="managing-tasks"
-        page="scanning"
-        title={_('Help: Tasks')}
-      />
-      {capabilities.mayOp('run_wizard') && (
-        <IconMenu icon={<WizardIcon />}>
-          {capabilities.mayCreate('task') && (
-            <MenuEntry
-              data-testid="task-wizard-menu"
-              title={_('Task Wizard')}
-              onClick={onTaskWizardClick}
-            />
-          )}
-          {capabilities.mayCreate('task') && (
-            <MenuEntry
-              data-testid="advanced-task-wizard-menu"
-              title={_('Advanced Task Wizard')}
-              onClick={onAdvancedTaskWizardClick}
-            />
-          )}
-          {mayUseModifyTaskWizard && (
-            <MenuEntry
-              data-testid="modify-task-wizard-menu"
-              title={_('Modify Task Wizard')}
-              onClick={onModifyTaskWizardClick}
-            />
-          )}
-        </IconMenu>
-      )}
-
-      <NewIconMenu
-        onNewClick={onTaskCreateClick}
-        onNewContainerClick={onContainerTaskCreateClick}
-      />
-    </IconDivider>
-  );
-};
-
-ToolBarIcons.propTypes = {
-  onAdvancedTaskWizardClick: PropTypes.func.isRequired,
-  onContainerTaskCreateClick: PropTypes.func.isRequired,
-  onModifyTaskWizardClick: PropTypes.func.isRequired,
-  onTaskCreateClick: PropTypes.func.isRequired,
-  onTaskWizardClick: PropTypes.func.isRequired,
-};
 
 const Page = ({
   filter,
@@ -164,7 +98,7 @@ const Page = ({
             sectionIcon={<TaskIcon size="large" />}
             table={TaskTable}
             title={_('Tasks')}
-            toolBarIcons={ToolBarIcons}
+            toolBarIcons={TaskToolBarIcons}
             onAdvancedTaskWizardClick={advancedTaskWizard}
             onContainerTaskCreateClick={createContainer}
             onError={onError}
